Split out the object tracker code
